+++ frontend/src/main/resources/webapp/src/app/main/books/books-home/book-grid/book-grid.component.ts
@@ -1,0 +1,20 @@
+import {Component, EventEmitter, OnInit, Output} from '@angular/core';
+
+@Component({
+  selector: 'app-book-grid',
+  templateUrl: './book-grid.component.html',
+  styleUrls: ['./book-grid.component.scss']
+})
+export class BookGridComponent implements OnInit {
+
+  @Output() switchMode = new EventEmitter<boolean>();
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+  switchToTable(){
+    this.switchMode.emit(true)
+  }
+}