+++ frontend/src/main/resources/webapp/src/app/main/books/books-home/books-home.component.ts
@@ -1,15 +1,21 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 
 @Component({
-  selector: 'app-books-home',
-  templateUrl: './books-home.component.html',
-  styleUrls: ['./books-home.component.scss']
+    selector: 'app-books-home',
+    templateUrl: './books-home.component.html',
+    styleUrls: ['./books-home.component.scss']
 })
 export class BooksHomeComponent implements OnInit {
 
-  constructor() { }
+    tableMode:boolean = false;
 
-  ngOnInit() {
-  }
+    constructor() {
+    }
 
+    ngOnInit() {
+    }
+
+    switchMode() {
+        this.tableMode = !this.tableMode;
+    }
 }