+++ frontend/src/main/resources/webapp/src/app/main/genres/genres-home/genres-home.component.ts
@@ -1,13 +1,27 @@
-import { Component, OnInit } from '@angular/core';
+import {AfterViewInit, Component, OnInit, ViewChild} from '@angular/core';
+import {OTableButtonComponent} from "ontimize-web-ngx";
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-genre-home',
   templateUrl: './genres-home.component.html',
   styleUrls: ['./genres-home.component.scss']
 })
-export class GenresHomeComponent implements OnInit {
+export class GenresHomeComponent implements OnInit, AfterViewInit {
 
-  constructor() { }
+  @ViewChild('genre_top')
+  protected genre_top: OTableButtonComponent;
+
+  ngAfterViewInit() {
+    this.genre_top.onClick.subscribe(event => {
+      this.router.navigate(['/main','genres','top'])
+    });
+  }
+
+
+  constructor(private router: Router) {
+
+  }
 
   ngOnInit() {
   }