+++ frontend/src/main/resources/webapp/src/app/main/genres/genres.module.ts
@@ -1,18 +1,26 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
 
-import { GenresRoutingModule } from './genres-routing.module';
-import { GenresHomeComponent } from './genres-home/genres-home.component';
+import {GenresRoutingModule} from './genres-routing.module';
+import {GenresHomeComponent} from './genres-home/genres-home.component';
 import {OntimizeWebModule} from "ontimize-web-ngx";
+import {GenresDetailComponent} from './genres-detail/genres-detail.component';
+import {SharedModule} from "../../shared/shared.module";
+import {GenresTopComponent} from "./genres-top/genres-top.component";
+import {BooksModule} from "../books/books.module";
 
 @NgModule({
-    imports: [
-        CommonModule,
-        GenresRoutingModule,
-        OntimizeWebModule
-    ],
+  imports: [
+    CommonModule,
+    GenresRoutingModule,
+    OntimizeWebModule,
+    SharedModule,
+    BooksModule
+  ],
   declarations: [
-      GenresHomeComponent
+      GenresHomeComponent,
+      GenresDetailComponent,
+      GenresTopComponent
   ]
 })
 export class GenresModule { }