+++ frontend/src/main/resources/webapp/src/app/main/main.component.ts
@@ -1,4 +1,5 @@
 import {Component, OnInit} from '@angular/core';
+import {LoginService} from "ontimize-web-ngx";
 
 @Component({
     selector: 'app-main',
@@ -7,10 +8,17 @@ import {Component, OnInit} from '@angular/core';
 })
 export class MainComponent implements OnInit {
 
-    constructor() {
+    user = "USERNAME";
+
+    constructor(public loginService: LoginService) {
     }
 
     ngOnInit() {
+        this.user = this.loginService.getSessionInfo().user
     }
 
+
+    logout() {
+        this.loginService.logoutWithConfirmationAndRedirect()
+    }
 }