+++ frontend/src/main/resources/webapp/src/app/main/publishers/publishers.module.ts
@@ -1,18 +1,23 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+
 import { OntimizeWebModule } from "ontimize-web-ngx";
 import { PublishersRoutingModule } from './publishers-routing.module';
 import {PublishersHomeComponent} from "./publishers-home/publishers-home.component";
+import {PublishersDetailComponent} from "./publishers-detail/publishers-detail.component";
+import {SharedModule} from "../../shared/shared.module";
 
 @NgModule({
   imports: [
     CommonModule,
     OntimizeWebModule,
-    PublishersRoutingModule
+    PublishersRoutingModule,
+      SharedModule
   ],
 
   declarations: [
-      PublishersHomeComponent
+      PublishersHomeComponent,
+      PublishersDetailComponent
   ]
 })
 export class PublishersModule { }